services/evaluator/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
anthropic==0.39.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
